# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/consultants": {
      "get": {
        "operationId": "getConsultants",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Column to sort by. Defaults to most recently created.",
            "schema": {
              "type": "string",
              "enum": [
                "displayId",
                "fullName",
                "email",
                "createdAt"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Free-text search across fullName, email and displayId",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": false,
            "in": "query",
            "description": "Filter by role ID (exact match)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleName",
            "required": false,
            "in": "query",
            "description": "Filter by role name (exact, case-insensitive)",
            "schema": {
              "example": "manager",
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filter by active flag",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated consultants",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedConsultantsEntity"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List consultants (paginated, filterable, sortable) — admin/manager only",
        "tags": [
          "Consultants"
        ]
      },
      "post": {
        "operationId": "createConsultant",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsultantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Consultant created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultantEntity"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new consultant",
        "tags": [
          "Consultants"
        ]
      }
    },
    "/consultants/me": {
      "get": {
        "operationId": "getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current consultant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultantEntity"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get my own consultant profile (any authenticated user)",
        "tags": [
          "Consultants"
        ]
      },
      "patch": {
        "operationId": "updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultantEntity"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update my own name (cannot change role or active status)",
        "tags": [
          "Consultants"
        ]
      }
    },
    "/consultants/by-display-id/{displayId}": {
      "get": {
        "operationId": "getConsultantByDisplayId",
        "parameters": [
          {
            "name": "displayId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Consultant found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultantEntity"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get consultant by display ID (consultant-XXXX)",
        "tags": [
          "Consultants"
        ]
      }
    },
    "/consultants/{id}": {
      "get": {
        "operationId": "getConsultant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Consultant found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultantEntity"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get consultant by ID",
        "tags": [
          "Consultants"
        ]
      },
      "patch": {
        "operationId": "updateConsultant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConsultantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Consultant updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultantEntity"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a consultant",
        "tags": [
          "Consultants"
        ]
      },
      "delete": {
        "operationId": "deleteConsultant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Consultant deactivated"
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deactivate a consultant (soft — sets isActive=false, blocks login)",
        "tags": [
          "Consultants"
        ]
      }
    },
    "/consultants/{id}/industries": {
      "put": {
        "operationId": "setConsultantIndustries",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetConsultantIndustriesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Industries updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultantEntity"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Replace a consultant's assigned industries (admin/manager only; self/peer-manager escalation rules enforced in the service)",
        "tags": [
          "Consultants"
        ]
      }
    },
...
  },
  "info": {
    "title": "Linktal Recruitment API",
    "description": "API for Linktal Recruitment OS",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RoleSummaryEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "consultant"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "ConsultantEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "displayId": {
            "type": "string",
            "example": "consultant-0001"
          },
          "azureId": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "example": "jane@linktal.com"
          },
          "fullName": {
            "type": "string",
            "example": "Jane Doe"
          },
          "roleId": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "role": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/RoleSummaryEntity"
              }
            ]
          },
          "industries": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "industryIds": {
            "description": "Industry IDs backing `industries`",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "displayId",
          "azureId",
          "email",
          "fullName",
          "roleId",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedConsultantsEntity": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConsultantEntity"
            }
          },
          "total": {
            "type": "number",
            "description": "Total rows matching the filters",
            "example": 16
          },
          "page": {
            "type": "number",
            "description": "Current page (1-based)",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "description": "Items per page",
            "example": 20
          },
          "pageCount": {
            "type": "number",
            "description": "Total number of pages",
            "example": 1
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "pageSize",
          "pageCount"
        ]
      },
      "UpdateMeDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "description": "Your display name",
            "example": "Jane Doe"
          }
        },
        "required": [
          "fullName"
        ]
      },
      "CreateConsultantDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "jane@linktal.com"
          },
          "fullName": {
            "type": "string",
            "description": "Full name",
            "example": "Jane Doe"
          },
          "roleId": {
            "type": "string",
            "description": "Role ID to assign (from the Role table)"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the consultant is active",
            "default": true
          }
        },
        "required": [
          "email",
          "fullName"
        ]
      },
      "UpdateConsultantDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "jane@linktal.com"
          },
          "fullName": {
            "type": "string",
            "description": "Full name",
            "example": "Jane Doe"
          },
          "roleId": {
            "type": "string",
            "description": "Role ID to assign (from the Role table)"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the consultant is active",
            "default": true
          },
          "roleName": {
            "type": "string",
            "enum": [
              "admin",
              "manager",
              "consultant",
              "finance",
              "researcher",
              "viewer"
            ]
          }
        }
      },
      "SetConsultantIndustriesDto": {
        "type": "object",
        "properties": {
          "industryIds": {
            "description": "Industry IDs (see /industries) — replaces the whole set",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "industryIds"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "code": {
            "type": "string",
            "example": "NOT_FOUND",
            "description": "Stable machine-readable code"
          },
          "message": {
            "type": "string",
            "example": "Candidate abc not found"
          },
          "details": {
            "nullable": true,
            "description": "Field-level messages (e.g. validation errors)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "path": {
            "type": "string",
            "example": "/api/candidates/abc"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-07-09T10:20:30.000Z"
          }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "details",
          "path",
          "timestamp"
        ]
      }
    }
  }
}